Refactored Rectificator and surrounding classes even more
