+++ lib/VideoStream.py
@@ -22,7 +22,6 @@ class VideoStream():
     def __init__(self, videoFilepath):
         self._vidcap = cv2.VideoCapture(videoFilepath)
         self.__imagesCache = pylru.lrucache(4) #set the size of cache to be 10 images large
-        self.__is_undistorted = True
 
         print("cv2 version", cv2.__version__)
         print ("num_of_frames", self.num_of_frames())
@@ -46,33 +45,35 @@ class VideoStream():
         # type: () -> int
         return int(self._vidcap.get(cv2.CAP_PROP_FPS))
 
-    def readImage(self, frameID):
+    def read_image(self, frameID):
         # type: (int) -> np
         if  frameID not in self.__imagesCache:
             # image is not in the cache. Read it from VideoCapture and save into cache
-            image = self.readFromVideoCapture(frameID)
+            image = self._read_image_raw(frameID)
             self.__imagesCache[frameID] = image
 
         return self.__imagesCache[frameID].copy()
 
-    def readImageObj(self, frameID):
+    def read_image_obj(self, frameID):
         # type: () -> Image
-        return Image(self.readImage(frameID))
+        return Image(self.read_image(frameID))
 
-    def readFromVideoCapture(self, frameID):
-        # type: (int) -> np
+    def read_image_undistorted(self, frameID: int) -> np:
+        image = self._read_image_raw(frameID)
+        image = Camera().undistortImage(image)
+        # Uncomment if need to show raw image
+        show_img = cv2.resize(image, (720, 576))
+        cv2.imshow('Debug', show_img)
+        cv2.waitKey(10)
+
+        return image
+
+    def _read_image_raw(self, frameID: int):
         self._vidcap.set(cv2.CAP_PROP_POS_FRAMES, float(frameID))
         success, image = self._vidcap.read()
         if not success:
             errorMessage = "Could not read frame " + str(frameID) + " from videofile"
             raise VideoStreamException(errorMessage)
-        if self.__is_undistorted:
-            # image = self.undistortImage(image, crop=cropped)
-            image = Camera().undistortImage(image)
-            # Uncomment if need to show raw image
-            show_img = cv2.resize(image, (720, 576))
-            cv2.imshow('Debug', show_img)
-            cv2.waitKey(10)
         return image
 
     def printMemoryUsage(self):
@@ -92,25 +93,10 @@ class VideoStream():
 
             try:
                 print ("get_id_of_first_frame", frame_id)
-                self.readFromVideoCapture(frame_id)
+                self._read_image_raw(frame_id)
                 return frame_id
             except VideoStreamException as error:
                 print("Cannot read frame " + str(frame_id) + ", skipping to next")
 
             frame_id += step_size
 
-    def setUndistorted(self, value: bool):
-        self.__is_undistorted = value
-
-    def setCropped(self, value: bool):
-        self.__is_cropped = value
-
-    def setUndistortDefault(self):
-        self.__is_undistorted = True
-        self.__is_cropped = True
-
-    def getCalibrationMatrix(self):
-        return self.__mtx
-
-    def getDistortionCoefficients(self):
-        return self.__dst
